+++ changelog.md
@@ -5,6 +5,7 @@
 > [!IMPORTANT]
 > This version is not released yet and is under active development.
 
+- Add a new `--no-config` option to disable loading of configuration files. Closes {issue}`750`.
 - Replace unmaintained `mergedeep` dependency by `deepmerge`.
 
 ## [5.1.1 (2025-08-24)](https://github.com/kdeldycke/click-extra/compare/v5.1.0...v5.1.1)